Typo in type #1067 (#1069)

+++ tensorflow_addons/layers/normalizations.py
@@ -75,7 +75,7 @@ class GroupNormalization(tf.keras.layers.Layer):
         self,
         groups: int = 2,
         axis: int = -1,
-        epsilon: int = 1e-3,
+        epsilon: float = 1e-3,
         center: bool = True,
         scale: bool = True,
         beta_initializer: types.Initializer = "zeros",
